Existing tournament page added
