+++ pong-app/frontend/src/app.tsx
@@ -15,7 +15,8 @@ const ResetPasswordPage = lazy(() => import('./pages/unauthorised/reset-password
 const ChangePasswordPage = lazy(() => import('./pages/unauthorised/changePassword'));
 const PlayPage = lazy(() => import('./pages/game/playPage'));
 const LobbyPage = lazy(() => import('./pages/authorised/lobby'));
-const TournamentPage = lazy(() => import('./pages/authorised/tournament'));
+const TournamentLocalPage = lazy(() => import('./pages/authorised/tournament-local'));
+const TournamentRemotePage = lazy(() => import('./pages/authorised/tournament-remote'));
 const QuickmatchLocalPage = lazy(() => import('./pages/authorised/quickmatch-local'));
 const QuickmatchRemotePage = lazy(() => import('./pages/authorised/quickmatch-remote'));
 const AvatarPage = lazy(() => import('./shared/avatar'));
@@ -115,7 +116,8 @@ const App = () => {
               <Route path="/lobby" element={<LobbyPage />} />
               <Route path="/quickmatch-local" element={<QuickmatchLocalPage />} />
               <Route path="/quickmatch-remote" element={<QuickmatchRemotePage />} />
-              <Route path="/tournament" element={<TournamentPage />} />
+              <Route path="/tournament-local" element={<TournamentLocalPage />} />
+              <Route path="/tournament-remote" element={<TournamentRemotePage />} />
             </Route>
 
             {/* Catch all route */}
